+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.11](https://github.com/streamingfast/substreams-ethereum/releases/tag/v0.9.11)
+
+* Added conversion from `&pb::BigInt` to `substreams::scalar::BigInt`.
+
+* Added conversion from `&pb::BigInt` to `substreams::scalar::BigDecimal`.
+
 ## [0.9.10](https://github.com/streamingfast/substreams-ethereum/releases/tag/v0.9.10)
 
 * Re-generated the Rust Ethereum bindings with latest Firehose Ethereum Protobuf definitions.